move 3commas apis to a ThreeCommas package
