+++ api/threecommas/rest/bots.go
