+++ api/threecommas/rest/bots_test.go
