+++ api/threecommas/rest/dobjs/bot.go
