+++ api/threecommas/rest/rest_test.go
