+++ api/threecommas/utils.go
@@ -1,4 +1,4 @@
-package api
+package threecommas
 
 import (
 	"crypto/hmac"
