+++ api/threecommas/utils_test.go
@@ -1,4 +1,4 @@
-package api
+package threecommas
 
 import "testing"
