+++ cmd/commacloner/serve.go
@@ -4,14 +4,14 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/jslowik/commacloner/api/websockets/dobjs"
+	"github.com/jslowik/commacloner/api/threecommas/websockets"
+	dobjs2 "github.com/jslowik/commacloner/api/threecommas/websockets/dobjs"
 	"io/ioutil"
 	"os"
 	"os/signal"
 	"time"
 
 	"github.com/gorilla/websocket"
-	"github.com/jslowik/commacloner/api/websockets"
 	"go.uber.org/zap"
 
 	"github.com/ghodss/yaml"
@@ -87,7 +87,7 @@ func serve(args []string) error {
 		return err
 	}
 
-	messageOut := make(chan *dobjs.IdentifierMessage)
+	messageOut := make(chan *dobjs2.IdentifierMessage)
 	interrupt := make(chan os.Signal, 1)
 	signal.Notify(interrupt, os.Interrupt)
 
@@ -102,7 +102,7 @@ func serve(args []string) error {
 	done := make(chan struct{})
 	go func() {
 		defer close(done)
-		pong := dobjs.IdentifierMessage{Type: "pong"}
+		pong := dobjs2.IdentifierMessage{Type: "pong"}
 
 		for {
 			msgType, message, readErr := conn.ReadMessage()
@@ -120,8 +120,8 @@ func serve(args []string) error {
 			}
 			logger.Debugf("recv: type - %d message - %s", msgType, message)
 
-			ctrlMessage := dobjs.Message{}
-			pingMessage := dobjs.PingMessage{}
+			ctrlMessage := dobjs2.Message{}
+			pingMessage := dobjs2.PingMessage{}
 			if unmarshalError := json.Unmarshal(message, &ctrlMessage); unmarshalError == nil {
 				switch ctrlMessage.Type {
 				case "welcome":
@@ -131,7 +131,7 @@ func serve(args []string) error {
 					logger.Infof("subscription confirmed : %s", message)
 				case "Deal", "Deal::ShortDeal":
 					logger.Debugf("received deal %v", ctrlMessage.Message)
-					dealMessage := dobjs.DealsMessage{}
+					dealMessage := dobjs2.DealsMessage{}
 					var dealErr error
 					if dealErr = json.Unmarshal(message, &dealMessage); dealErr == nil {
 						dealErr = stream.HandleDeal(dealMessage)
